refactor: reorganize cycle directory
