+++ autora/cycle/result.py
