--- autora/cycle/result/serializer/__init__.py
@@ -1,12 +1,0 @@
-from typing import Protocol, runtime_checkable
-
-from autora.cycle.result import ResultCollection
-
-
-@runtime_checkable
-class ResultCollectionSerializer(Protocol):
-    def load(self) -> ResultCollection:
-        ...
-
-    def dump(self, data: ResultCollection) -> None:
-        ...
